Move misplaced practiceMode test and simplify integration test script
