+++ server/test/test_latest_mlp_model.py
@@ -14,8 +14,7 @@ SERVER_DIR = Path(__file__).resolve().parents[1]
 
 def _get_free_port() -> int:
     with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-    s.bind(("127.0.0.1", 0))
-
+        s.bind(("127.0.0.1", 0))
         return s.getsockname()[1]
